+++ src/ufuncs.c
@@ -2307,6 +2307,7 @@ static void from_wkb_func(char** args, npy_intp* dimensions, npy_intp* steps,
   char *ip1 = args[0], *ip2 = args[1], *op1 = args[2];
   npy_intp is1 = steps[0], is2 = steps[1], os1 = steps[2];
   PyObject* in1;
+  char ignore_invalid = *(npy_bool*)ip2;
   npy_intp n = dimensions[0];
   npy_intp i;
   GEOSWKBReader* reader;
@@ -2332,9 +2333,6 @@ static void from_wkb_func(char** args, npy_intp* dimensions, npy_intp* steps,
   for (i = 0; i < n; i++, ip1 += is1, op1 += os1) {
     /* ip1 is pointer to array element PyObject* */
     in1 = *(PyObject**)ip1;
-
-    /* ip2 is a pointer to a scalar bool */
-    char ignore_invalid = *(npy_bool*)ip2;
 
     if (in1 == Py_None) {
       /* None in the input propagates to the output */
@@ -2374,12 +2372,11 @@ static void from_wkb_func(char** args, npy_intp* dimensions, npy_intp* steps,
       }
       if (ret_ptr == NULL) {
         if (ignore_invalid) {
-            // the output geometry will be set to None
-            errstate = PGWARN_INVALID_WKB;
-        }
-        else {
-            errstate = PGERR_GEOS_EXCEPTION;
-            goto finish;
+          // the output geometry will be set to None
+          errstate = PGWARN_INVALID_WKB;
+        } else {
+          errstate = PGERR_GEOS_EXCEPTION;
+          goto finish;
         }
       }
     }
@@ -2392,14 +2389,23 @@ finish:
 }
 static PyUFuncGenericFunction from_wkb_funcs[1] = {&from_wkb_func};
 
-static char from_wkt_dtypes[2] = {NPY_OBJECT, NPY_OBJECT};
+static char from_wkt_dtypes[3] = {NPY_OBJECT, NPY_BOOL, NPY_OBJECT};
 static void from_wkt_func(char** args, npy_intp* dimensions, npy_intp* steps,
                           void* data) {
-  GEOSGeometry* ret_ptr;
+  char *ip1 = args[0], *ip2 = args[1], *op1 = args[2];
+  npy_intp is1 = steps[0], is2 = steps[1], os1 = steps[2];
   PyObject* in1;
-
+  char ignore_invalid = *(npy_bool*)ip2;
+  npy_intp n = dimensions[0];
+  npy_intp i;
+  GEOSGeometry* ret_ptr;
   GEOSWKTReader* reader;
   const char* wkt;
+
+  if ((is2 != 0)) {
+    PyErr_Format(PyExc_ValueError, "from_wkt function called with non-scalar parameters");
+    return;
+  }
 
   GEOS_INIT;
 
@@ -2410,7 +2416,7 @@ static void from_wkt_func(char** args, npy_intp* dimensions, npy_intp* steps,
     goto finish;
   }
 
-  UNARY_LOOP {
+  for (i = 0; i < n; i++, ip1 += is1, op1 += os1) {
     /* ip1 is pointer to array element PyObject* */
     in1 = *(PyObject**)ip1;
 
@@ -2439,8 +2445,13 @@ static void from_wkt_func(char** args, npy_intp* dimensions, npy_intp* steps,
       /* Read the WKT */
       ret_ptr = GEOSWKTReader_read_r(ctx, reader, wkt);
       if (ret_ptr == NULL) {
-        errstate = PGERR_GEOS_EXCEPTION;
-        goto finish;
+        if (ignore_invalid) {
+          // the output geometry will be set to None
+          errstate = PGWARN_INVALID_WKT;
+        } else {
+          errstate = PGERR_GEOS_EXCEPTION;
+          goto finish;
+        }
       }
     }
     OUTPUT_Y;
@@ -2880,7 +2891,7 @@ int init_ufuncs(PyObject* m, PyObject* d) {
   DEFINE_GENERALIZED(create_collection, 2, "(i),()->()");
 
   DEFINE_CUSTOM(from_wkb, 2);
-  DEFINE_CUSTOM(from_wkt, 1);
+  DEFINE_CUSTOM(from_wkt, 2);
   DEFINE_CUSTOM(to_wkb, 5);
   DEFINE_CUSTOM(to_wkt, 5);
   DEFINE_CUSTOM(from_shapely, 1);